Revert "fix: modify some controllers and transfer some"
