+++ backend/route/manufacturerRoute.js
@@ -2,6 +2,7 @@ import express from "express";
 import * as manufacturerController from "../controller/manufacturerController.js";
 
 const router = express.Router({ mergeParams: true });
+router.get(":userId", manufacturerController.getAllManufacturers);
 router.get("/:id", manufacturerController.getManufacturerById);
 router.post("/", manufacturerController.createManufacturer);
 router.patch("/:id", manufacturerController.updateManufacturer);
