+++ backend/route/notificationRoute.js
@@ -2,6 +2,7 @@ import express from "express";
 import * as notificationController from "../controller/notificationController.js";
 
 const router = express.Router({ mergeParams: true });
+router.get("/:userId", notificationController.getAllNotifications);
 router.get("/:id", notificationController.getNotificationById);
 router.post("/", notificationController.createNotification);
 router.patch("/:id", notificationController.updateNotification);
