+++ frontend/src/page/inventory/Inventory.jsx
@@ -1,27 +1,7 @@
-import React, { useState, useEffect} from "react";
-import axios from "axios";
+import React from "react";
 
 const Inventory = () => {
-
-  const [inventory, setInventory] = useState("");
-
-  useEffect(() => {
-    axios.get("http://localhost:5555/user/:66654dc4c6e950671e988962/inv")
-    .then((res) => {
-        console.log(res)
-        setInventory(res.data.data.docs);
-        console.log(inventory);
-    })
-    .catch((err) => {
-        console.error(err);
-    })
-  })
-
-  return (
-    <div>
-      <p>{ inventory.inventories[0] }</p>
-    </div>
-  )
+  return <div>Inventory</div>;
 };
 
 export default Inventory;
